requirements.txt
python-telegram-bot
eyed3
mutagen
requests
soundcloud
sacad